refactor: split Lambda by responsibility
