+++ infra/aws/lambda.ts
@@ -16,18 +16,24 @@ import { sesDomainIdentity } from './ses';
 const currentRegion = aws.getRegion();
 const currentIdentity = aws.getCallerIdentity();
 
-// Lambda function name (defined early for log group)
-const lambdaName = 'slackmail-juzumaru';
+// =============================================================================
+// S3 Lambda (Email Processing)
+// =============================================================================
 
-// CloudWatch Log Group for Lambda
-export const lambdaLogGroup = new aws.cloudwatch.LogGroup('lambda-log-group', {
-  name: `/aws/lambda/${lambdaName}`,
-  retentionInDays: 14,
-  tags,
-});
+const s3LambdaName = 'slackmail-s3-handler';
 
-// IAM role for Lambda execution
-export const lambdaRole = new aws.iam.Role('lambda-role', {
+// CloudWatch Log Group for S3 Lambda
+export const s3LambdaLogGroup = new aws.cloudwatch.LogGroup(
+  's3-lambda-log-group',
+  {
+    name: `/aws/lambda/${s3LambdaName}`,
+    retentionInDays: 14,
+    tags,
+  },
+);
+
+// IAM role for S3 Lambda execution
+export const s3LambdaRole = new aws.iam.Role('s3-lambda-role', {
   assumeRolePolicy: JSON.stringify({
     Version: '2012-10-17',
     Statement: [
@@ -43,28 +49,31 @@ export const lambdaRole = new aws.iam.Role('lambda-role', {
   tags,
 });
 
-// Policy: CloudWatch Logs access (scoped to specific log group)
-export const lambdaLogsPolicy = new aws.iam.RolePolicy('lambda-logs-policy', {
-  role: lambdaRole.id,
-  policy: pulumi
-    .all([currentRegion, currentIdentity])
-    .apply(([region, identity]) =>
-      JSON.stringify({
-        Version: '2012-10-17',
-        Statement: [
-          {
-            Effect: 'Allow',
-            Action: ['logs:CreateLogStream', 'logs:PutLogEvents'],
-            Resource: `arn:aws:logs:${region.name}:${identity.accountId}:log-group:/aws/lambda/${lambdaName}:*`,
-          },
-        ],
-      }),
-    ),
-});
+// Policy: CloudWatch Logs access for S3 Lambda
+export const s3LambdaLogsPolicy = new aws.iam.RolePolicy(
+  's3-lambda-logs-policy',
+  {
+    role: s3LambdaRole.id,
+    policy: pulumi
+      .all([currentRegion, currentIdentity])
+      .apply(([region, identity]) =>
+        JSON.stringify({
+          Version: '2012-10-17',
+          Statement: [
+            {
+              Effect: 'Allow',
+              Action: ['logs:CreateLogStream', 'logs:PutLogEvents'],
+              Resource: `arn:aws:logs:${region.name}:${identity.accountId}:log-group:/aws/lambda/${s3LambdaName}:*`,
+            },
+          ],
+        }),
+      ),
+  },
+);
 
 // Policy: S3 read access for email bucket
-export const lambdaS3Policy = new aws.iam.RolePolicy('lambda-s3-policy', {
-  role: lambdaRole.id,
+export const s3LambdaS3Policy = new aws.iam.RolePolicy('s3-lambda-s3-policy', {
+  role: s3LambdaRole.id,
   policy: pulumi.interpolate`{
     "Version": "2012-10-17",
     "Statement": [
@@ -79,10 +88,91 @@ export const lambdaS3Policy = new aws.iam.RolePolicy('lambda-s3-policy', {
   }`,
 });
 
-// Policy: SES send email access
-export const lambdaSesPolicy = new aws.iam.RolePolicy('lambda-ses-policy', {
-  role: lambdaRole.id,
-  policy: pulumi.interpolate`{
+// S3 Lambda function (email processing)
+export const s3Lambda = new aws.lambda.Function('s3-lambda', {
+  name: s3LambdaName,
+  runtime: aws.lambda.Runtime.NodeJS22dX,
+  handler: 's3-handler.handler',
+  role: s3LambdaRole.arn,
+  code: new pulumi.asset.AssetArchive({
+    's3-handler.js': new pulumi.asset.FileAsset('./dist/s3-handler.js'),
+  }),
+  timeout: 30,
+  memorySize: 256,
+  environment: {
+    variables: {
+      NODE_ENV: stackName,
+      EMAIL_BUCKET_NAME: emailBucket.bucket,
+      SLACK_SIGNING_SECRET: slackSigningSecret,
+      SLACK_BOT_TOKEN: slackBotToken,
+      SLACK_CHANNEL_ID: slackChannelId,
+      SENTRY_DSN: sentryDsn,
+    },
+  },
+  tags,
+});
+
+// =============================================================================
+// Slack Lambda (API Gateway / Slack Events)
+// =============================================================================
+
+const slackLambdaName = 'slackmail-slack-handler';
+
+// CloudWatch Log Group for Slack Lambda
+export const slackLambdaLogGroup = new aws.cloudwatch.LogGroup(
+  'slack-lambda-log-group',
+  {
+    name: `/aws/lambda/${slackLambdaName}`,
+    retentionInDays: 14,
+    tags,
+  },
+);
+
+// IAM role for Slack Lambda execution
+export const slackLambdaRole = new aws.iam.Role('slack-lambda-role', {
+  assumeRolePolicy: JSON.stringify({
+    Version: '2012-10-17',
+    Statement: [
+      {
+        Action: 'sts:AssumeRole',
+        Principal: {
+          Service: 'lambda.amazonaws.com',
+        },
+        Effect: 'Allow',
+      },
+    ],
+  }),
+  tags,
+});
+
+// Policy: CloudWatch Logs access for Slack Lambda
+export const slackLambdaLogsPolicy = new aws.iam.RolePolicy(
+  'slack-lambda-logs-policy',
+  {
+    role: slackLambdaRole.id,
+    policy: pulumi
+      .all([currentRegion, currentIdentity])
+      .apply(([region, identity]) =>
+        JSON.stringify({
+          Version: '2012-10-17',
+          Statement: [
+            {
+              Effect: 'Allow',
+              Action: ['logs:CreateLogStream', 'logs:PutLogEvents'],
+              Resource: `arn:aws:logs:${region.name}:${identity.accountId}:log-group:/aws/lambda/${slackLambdaName}:*`,
+            },
+          ],
+        }),
+      ),
+  },
+);
+
+// Policy: SES send email access for Slack Lambda
+export const slackLambdaSesPolicy = new aws.iam.RolePolicy(
+  'slack-lambda-ses-policy',
+  {
+    role: slackLambdaRole.id,
+    policy: pulumi.interpolate`{
     "Version": "2012-10-17",
     "Statement": [
       {
@@ -95,30 +185,37 @@ export const lambdaSesPolicy = new aws.iam.RolePolicy('lambda-ses-policy', {
       }
     ]
   }`,
-});
+  },
+);
 
-// Lambda function
-export const boltLambda = new aws.lambda.Function('slackmail-juzumaru', {
-  name: lambdaName,
+// Slack Lambda function (API Gateway / Slack events)
+export const slackLambda = new aws.lambda.Function('slack-lambda', {
+  name: slackLambdaName,
   runtime: aws.lambda.Runtime.NodeJS22dX,
-  handler: 'index.handler',
-  role: lambdaRole.arn,
+  handler: 'slack-handler.handler',
+  role: slackLambdaRole.arn,
   code: new pulumi.asset.AssetArchive({
-    // Bundled by esbuild (pnpm build:lambda)
-    'index.js': new pulumi.asset.FileAsset('./dist/index.js'),
+    'slack-handler.js': new pulumi.asset.FileAsset('./dist/slack-handler.js'),
   }),
   timeout: 30,
   memorySize: 256,
   environment: {
     variables: {
       NODE_ENV: stackName,
-      EMAIL_BUCKET_NAME: emailBucket.bucket,
       EMAIL_DOMAIN: emailDomain,
       SLACK_SIGNING_SECRET: slackSigningSecret,
       SLACK_BOT_TOKEN: slackBotToken,
       SLACK_CHANNEL_ID: slackChannelId,
-      SENTRY_DSN: sentryDsn,
     },
   },
   tags,
 });
+
+// =============================================================================
+// Backwards compatibility exports
+// =============================================================================
+
+// Export for backwards compatibility (used by existing code)
+export const boltLambda = slackLambda;
+export const lambdaRole = s3LambdaRole;
+export const lambdaLogGroup = s3LambdaLogGroup;
